requirements.txt
PyGithub
rich
psutil
python-dotenv
